+++ src/components/incomes/IncomeList.jsx
@@ -3,6 +3,7 @@ import useDeleteTransaction from "../../hooks/useDeleteTransaction";
 import { useGetTransactions } from "../../hooks/useGetTransactions";
 import styles from "./IncomeList.module.css";
 import EditIncomePopup from "./EditIncomePopup";
+import FeedbackPopup from "../expenses/FeedbackPopup";
 
 function formatAmount(amount) {
   return new Intl.NumberFormat("en-NP", {
@@ -16,12 +17,26 @@ function IncomeList({ incomes }) {
   const { loading } = useGetTransactions();
   const [showEditPopup, setShowEditPopup] = useState(false);
   const [transactionToEdit, setTransactionToEdit] = useState(null);
+  const [feedbackMessage, setFeedbackMessage] = useState("");
+  const [isError, setIsError] = useState(false);
 
   const handleDeleteIncome = async (transactionId) => {
+    const isConfirmed = window.confirm(
+      "Are you sure you want to delete this expense?"
+    );
+    if (!isConfirmed) {
+      return;
+    }
     try {
       await deleteTransaction("incomes", transactionId);
+      setFeedbackMessage("Income transaction deleted successfully.");
+      setIsError(false);
+      setTimeout(() => setFeedbackMessage(""), 3000);
     } catch (error) {
       console.error("Error deleting income:", error);
+      setFeedbackMessage("Error deleting income. Please try again.");
+      setIsError(true);
+      setTimeout(() => setFeedbackMessage(""), 3000);
     }
   };
 
@@ -32,6 +47,9 @@ function IncomeList({ incomes }) {
 
   const closeEditPopup = () => {
     setShowEditPopup(false);
+  };
+  const closeFeedbackPopup = () => {
+    setFeedbackMessage("");
   };
 
   return (
@@ -47,6 +65,7 @@ function IncomeList({ incomes }) {
         <table className={styles.incomeTable}>
           <thead>
             <tr>
+              <th>SN</th>
               <th>Amount</th>
               <th>Category</th>
               <th>Description</th>
@@ -57,9 +76,10 @@ function IncomeList({ incomes }) {
           <tbody>
             {incomes.map((income, index) => (
               <tr key={index}>
+                <td>{index + 1}</td>
                 <td>{formatAmount(income.transactionAmount)}</td>
                 <td>{income.categoryName}</td>
-                <td>{income.description}</td>
+                <td className={styles.description}>{income.description}</td>
                 <td>{income.date}</td>
                 <td>
                   <div className={styles.btns}>
@@ -88,6 +108,9 @@ function IncomeList({ incomes }) {
           onClose={closeEditPopup}
           transaction={transactionToEdit}
         />
+      )}
+      {feedbackMessage && (
+        <FeedbackPopup message={feedbackMessage} onClose={closeFeedbackPopup} />
       )}
     </div>
   );
